tast-tests: add runtime var crash.markTestInProgress/testInProgressPrefix
crash.markTestInProgress controls if create file test-in-prog.
crash.testInProgressPrefix allow a customized prefix be writen to test-in-prog.

BUG=b:193461265

TEST=tast run -var=crash.prefix=cuj. <ip> example.RuntimeVars
TEST=tast run -var=crash.testInProgFile=false  <ip> example.RuntimeVars
TEST=fast_build.sh -T

+++ src/chromiumos/tast/local/crash/crash.go
@@ -14,6 +14,7 @@ import (
 	"path/filepath"
 	"regexp"
 	"sort"
+	"strconv"
 	"strings"
 	"time"
 
@@ -109,6 +110,20 @@ const (
 	// FilterInIgnoreAllCrashes is a value to put in the filter-in file if
 	// you wish to ignore all crashes that happen during a test.
 	FilterInIgnoreAllCrashes = "none"
+)
+
+var (
+	markTestInProgressVar = testing.RegisterVarString(
+		"crash.markTestInProgress", // The variable controls if the test-in-prog file will be created.
+		// Default value is true, create file test-in-prog by default.
+		// When set the var to "false", test-in-prog file will not be created.
+		"true",
+		"The variable that controls if test-in-prog file will be created")
+
+	testInProgressPrefixVar = testing.RegisterVarString(
+		"crash.testInProgressPrefix", // The variable prefixed to the test case name in test-in-prog file.
+		"",                           // By default no prefix will be added.
+		"the string that will be prefixed to the test case name in test-in-prog file.")
 )
 
 // DefaultDirs returns all standard directories to which crashes are written.
@@ -485,17 +500,37 @@ func processRunning(procName string) (bool, error) {
 	return false, nil
 }
 
+// shouldMarkTestInProgress Parses markTestInProgressVar to a boolean value.
+// If the value of markTestInProgressVar is not supported by strconv.ParseBool(), it will return true.
+func shouldMarkTestInProgress(ctx context.Context) bool {
+	testInProgFile, err := strconv.ParseBool(markTestInProgressVar.Value())
+
+	//If any parse error happens, set the value to true.
+	if err != nil {
+		testing.ContextLogf(ctx, "Failed to parse %q, using default value true", markTestInProgressVar.Value())
+		testInProgFile = true
+	}
+	return testInProgFile
+}
+
 // MarkTestInProgress writes |name| to |testInProgressPath|, indicating to crash_reporter
 // that the given test is in progress.
-func MarkTestInProgress(name string) error {
-	if err := ioutil.WriteFile(testInProgressPath, []byte(name), 0644); err != nil {
+func MarkTestInProgress(ctx context.Context, name string) error {
+	if !shouldMarkTestInProgress(ctx) {
+		return nil
+	}
+	if err := ioutil.WriteFile(testInProgressPath, []byte(testInProgressPrefixVar.Value()+name), 0644); err != nil {
 		return errors.Wrap(err, "failed to write in-progress test name")
 	}
 	return nil
 }
 
 // MarkTestDone removes the file indicating which test is running.
-func MarkTestDone() error {
+func MarkTestDone(ctx context.Context) error {
+	// Don't remove the test-in-prog file if it's not created by tast in MarkTestInProgress function.
+	if !shouldMarkTestInProgress(ctx) {
+		return nil
+	}
 	if err := os.Remove(testInProgressPath); err != nil && !os.IsNotExist(err) {
 		return errors.Wrap(err, "failed to remove in-progress test name")
 	}
